Define the WebContent process as launched in WebPage::create https://bugs.webkit.org/show_bug.cgi?id=257964 rdar://110648262
Reviewed by Brent Fulgham.

From a sandbox perspective, define the WebContent process as launched in WebPage::create. This ensures that the
sandbox state variable is set before any content is loaded, which is not immediately clear from doing it when
the load is committed.

* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::create):
(WebKit::WebPage::didCommitLoad):

Canonical link: https://commits.webkit.org/265080@main