REGRESSION (262850@main): [ iOS, macOS Debug ] TestWebKitAPI.IPCTestingAPI.CanInterceptHasStorageAccess is a constant failure. https://bugs.webkit.org/show_bug.cgi?id=255563 rdar://108168809
Reviewed by NOBODY (OOPS!).

In 262850@main, I modified the jsValueForDecodedArgumentValue() taking an
ObjectIdentifier<U> to take an ObjectIdentifierGenericBase instead, in order
to support AtomicObjectIdentifiers as well. However, this broke the API test,
indicating that the wrong jsValueForDecodedArgumentValue() is likely getting
picked now. To address the issue, I am restoring the ObjectIdentifier<U>
overload and introducing a new AtomicObjectIdentifier<U> instead.

* Source/WebKit/Platform/IPC/JSIPCBinding.h:
(IPC::jsValueForDecodedArgumentValue):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/IPCTestingAPI.mm: