[JSC] Add more assertions to help diagnose a crash below JSModuleRecord::instantiateDeclarations https://bugs.webkit.org/show_bug.cgi?id=257896 rdar://110339520
Reviewed by NOBODY (OOPS!).

Previously, we landed a patch [1] to mitigate an unexplained crash
due to an undefined JSModuleEnvironment* in JSModuleRecord::instantiateDeclarations.
Initially, we suspected the crash due to an undefined entry gotten from
the dependencies map. With info collected by the follow-up patch [2], we found that the
entry is not undefined but with an undefined property `module`. This might
be due to the potential module instantiation failure before filling the registered
entry in requestInstantiate of ModuleLoader.js. To support our hypothesis,
add more assertions for more info in AbstractModuleRecord::hostResolveImportedModule.

[1] https://commits.webkit.org/262608@main
[2] https://commits.webkit.org/262687@main

* Source/JavaScriptCore/runtime/AbstractModuleRecord.cpp:
(JSC::AbstractModuleRecord::hostResolveImportedModule):
* Source/JavaScriptCore/runtime/JSCJSValueInlines.h:
(JSC::JSValue::getPropertySlot const):