[iOS] Allow syscall-unix SYS_faccessat in the GPU Process sandbox (Part 2) https://bugs.webkit.org/show_bug.cgi?id=257985 <rdar://108203015>
Reviewed by NOBODY (OOPS!).

We have discovered that some aspects of Metal's shader caching needs to use
SYS_faccessat to check for file presence. When blocked by the Sandbox it
immediately fails, turning off the binary archive capability.

This patch restores the feature. I will work to determine whether the small
memory regression we measured after my last attempt was a true memory regression,
or simply reverting a false memory progression created when we blocked it in
our April 2023 releases.

* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.sb.in: