Do not create blob temp folder for ephemeral sessions https://bugs.webkit.org/show_bug.cgi?id=257950 rdar://110545873
Reviewed by Sihui Liu.

We are creating a temp folder for blobs to use mapped memory to store blobs, as a way to limit networking process jetsams.
This creation was not dependent on session being ephemeral or not.
We are now skipping the creation of this temp folder for ephemeral sessions.
This will trigger use of regular memory for storing blobs, which may increase networking process jetsams.

* Source/WebKit/NetworkProcess/cocoa/NetworkSessionCocoa.mm:
(WebKit::NetworkSessionCocoa::NetworkSessionCocoa):

Canonical link: https://commits.webkit.org/265127@main