Fix termination for ApplyStyleCommand::applyRelativeFontStyleChange https://bugs.webkit.org/show_bug.cgi?id=258145 rdar://110319440
Reviewed by Ryosuke Niwa.

This change fixes applyRelativeFontStyleChange in case where a node is
removed when we try to add a surrounding span, but we cannot add the
node back as removal makes it so that the surrounding span isn't
editable. The fix ensures that we terminate the loop when the node in
question doesn't have any parent.

* LayoutTests/fast/editing/editing-with-design-mode-crash-expected.txt: Added.
* LayoutTests/fast/editing/editing-with-design-mode-crash.html: Added.
* Source/WebCore/editing/ApplyStyleCommand.cpp:
(WebCore::ApplyStyleCommand::applyRelativeFontStyleChange):

Canonical link: https://commits.webkit.org/265226@main