[MQ4] Implement `update` media feature https://bugs.webkit.org/show_bug.cgi?id=180245 rdar://35799713
Reviewed by Darin Adler.

From https://drafts.csswg.org/mediaqueries/#update :
The update media feature is used to query the ability of the output device to modify the appearance of content once it has been rendered. It accepts the following values:

- none
Once it has been rendered, the layout can no longer be updated. Example: documents printed on paper.
- slow
The layout may change dynamically according to the usual rules of CSS, but the output device is not able to render or display changes quickly enough for them to be perceived as a smooth animation. Example: E-ink screens or severely under-powered devices.
- fast
The layout may change dynamically according to the usual rules of CSS, and the output device is not unusually constrained in speed, so regularly-updating things like CSS Animations can be used. Example: computer screens.

This is part of Interop 2023.

Also update overflow-block: paged to match on print mediums.

* LayoutTests/fast/css/media-query-overflow-block-paged-print-expected.html: Added.
* LayoutTests/fast/css/media-query-overflow-block-paged-print.html: Added.
* LayoutTests/fast/css/media-query-update-none-expected.html: Added.
* LayoutTests/fast/css/media-query-update-none.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/test_media_queries-expected.txt:
* Source/WebCore/css/query/MediaQueryFeatures.cpp:
(WebCore::MQ::Features::update):
(WebCore::MQ::Features::overflowBlock):
* Source/WebCore/css/query/MediaQueryFeatures.h:

Canonical link: https://commits.webkit.org/265277@main