REGRESSION(264826@main): [GStreamer] Fill information on the gst GL context https://bugs.webkit.org/show_bug.cgi?id=258199
Reviewed by Miguel Gomez.

In 264826@main I removed the activate and fill of the gst GL context,
confused by the API docs of gst_gl_context_fill_info() that says that
it's typically used with wrapped contexts to allow wrapped contexts to
be used as regular GstGLContexts. We don't want it to be used as a
regular context, but the regular context uses the filled information of
the sharing context to get the EGL config.

* Source/WebCore/platform/graphics/egl/GLContext.cpp:
(WebCore::GLContext::unmakeContextCurrent):
(WebCore::GLContext::ScopedGLContextCurrent::ScopedGLContextCurrent):
(WebCore::GLContext::ScopedGLContextCurrent::~ScopedGLContextCurrent):
* Source/WebCore/platform/graphics/egl/GLContext.h:
* Source/WebCore/platform/graphics/gstreamer/PlatformDisplayGStreamer.cpp:
(WebCore::PlatformDisplay::gstGLContext const):

Canonical link: https://commits.webkit.org/265302@main