WebProcess hung on main thread in WebPage::platformInitializeAccessibility https://bugs.webkit.org/show_bug.cgi?id=258140 rdar://110191499
Reviewed by Brent Fulgham.

The WebProcess can hang on the main thread in WebPage::platformInitializeAccessibility for several seconds.
This appears to be caused by the call to [NSApplication _accessibilityInitialize] trying to connect to the
Launch Services daemon. At this point, the daemon is blocked in the sandbox. Fix this by extending the time
holding the sandbox extension to the daemon, which has been revoked.

* Source/WTF/wtf/PlatformHave.h:
* Source/WebKit/WebProcess/WebPage/mac/WebPageMac.mm:
(WebKit::WebPage::platformInitializeAccessibility):
* Source/WebKit/WebProcess/WebProcess.h:
* Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::platformInitializeWebProcess):
(WebKit::WebProcess::revokeLaunchServicesSandboxExtension):
* Source/WebKit/WebProcess/cocoa/LaunchServicesDatabaseManager.h:
(WebKit::LaunchServicesDatabaseManager::hasReceivedLaunchServicesDatabase const):

Canonical link: https://commits.webkit.org/265334@main