REGRESSION(259684@main-259698@main?) [[ Ventura wk2 ] fast/images/avif-as-image.html is a constant ImageOnlyFailure https://bugs.webkit.org/show_bug.cgi?id=256119 rdar://108684696
Reviewed by Per Arne Vollan.

We had code of the form:
```
callIntoWebKitAdditionsToMakeImagesWork();
```

There's no reason for the "make images work" code to be in WebKitAdditions, so this patch
just moves it to WebKit and deletes the USE(APPLE_INTERNAL_SDK) guard. This fixes the images.

* LayoutTests/platform/mac/TestExpectations:
* Source/WebCore/PAL/PAL.xcodeproj/project.pbxproj:
* Source/WebCore/PAL/pal/cf/VideoToolboxSoftLink.cpp:
* Source/WebCore/PAL/pal/cf/VideoToolboxSoftLink.h:
* Source/WebCore/PAL/pal/spi/cf/VideoToolboxSPI.h: Added.
* Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::setVideoDecoderBehaviors):
(WebKit::WebProcess::platformInitializeWebProcess):

Canonical link: https://commits.webkit.org/265359@main