Remove unnecessary release assertion from mixed content checker. https://bugs.webkit.org/show_bug.cgi?id=258303 <rdar://110766912>
Reviewed by Brent Fulgham.

We now check the entire frame tree for mixed content checks for all
resources loads. An assertion that a document has a frame is no
longer valid in general. This assertion was originally added in
215749@main in an attempt to cover an untestable case. This replaces
the assert with a null check.

* Source/WebCore/loader/MixedContentChecker.cpp:
(WebCore::foundMixedContentInFrameTree):

Canonical link: https://commits.webkit.org/259548.841@safari-7615-branch