[WPE][Tools] cross-toolchain-helper: Add a RISCV64 target https://bugs.webkit.org/show_bug.cgi?id=257789
Reviewed by Philippe Normand.

This adds a qemu-riscv64 target that can be used to easily
cross-build for RISCV64. This will help with developing
the port. Once a popular RISCV64 board is out we can
replace the qemu target with that one.

To test this simply pass the flag --cross-target=qemu-riscv64
to the script build-webkit

* Tools/yocto/riscv/bblayers.conf: Added.
* Tools/yocto/riscv/fix-nativesdk-image-defs.patch: Added.
* Tools/yocto/riscv/local-qemuriscv64.conf: Added.
* Tools/yocto/riscv/manifest.xml: Added.
* Tools/yocto/targets.conf:

Canonical link: https://commits.webkit.org/265390@main